+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Rename GitCloner to GitHandler
 - restructure code in review command
 - code command does not require argument if executed in ticket directory
+- change to relative path outputs in log
 
 ### Removed